PasswordID int32 -> int64

+++ internal/entity/password.go
@@ -3,10 +3,10 @@ package entity
 import "time"
 
 type Password struct {
-	ID             int32      `json:"id"`
+	ID             int64      `json:"id"`
 	UserID         int64      `json:"userId"`
 	PasswordHash   string     `json:"passwordHash"`
-	FailedAttempts int32      `json:"failedAttempts"`
+	FailedAttempts int64      `json:"failedAttempts"`
 	CreatedAt      time.Time  `json:"createdAt"`
 	UpdatedAt      time.Time  `json:"updatedAt"`
 	DeletedAt      *time.Time `json:"deletedAt"`
